Revamps moving platform mechanics
Moving platform now has a boolean that determines whether it moves or
not. When initialized, the platform width can be adjusted to follow the
size that it needs. The platform has a start and end point that tells it
where to start and stop. The platform begins in the center of the two
points, and the width of the two start and end points are the distance
between the center of the platform. Specify 10 and the end point will be
10 units in the positive x direction from the platform and the start
point will be 10 units in the negative x direction.