Revamps lights
Lights now have a custom scale based on whether they have the torch on
them or not. When the player has the torch, the scale of the light is
bigger to improve visibility. The torch holders now affect the moving
platforms. To assign a torch holder to a certain platform, assign the
torch holder to the appropriate node2D of the platform. If the torch
affects the door instead, tick the boolean that in available in the
editor. That determines the purpose of the placeholder.