Adds the darkvision rune
The darkvision rune temporarily gives the player darkvision, allowing
them to see a majority of the room for a while before it goes back to
being dark. Once the player collides with the rune, it disappears and
the has_darkvision boolean in the player is made true, enlarging vision
size to the vision radius, then slowly reducing it by the
vision_reduction_speed until it goes back to a scale of 1.