Fixes trapdoor and lever
Trapdoor and lever not functioning properly previously. Now the trapdoor
has an is_activated bool that can be tracked. Lever can now call that
bool instead of running a function everytime. Connects the onbodyentered
and onbodyexited which wasn't previously connected, causing the trapdoor
to open and close when interacting with other things.